Precio de alquiler como float y corrección de que si el precio es cero no aparecía

+++ API/database/create-tables.sql
@@ -26,7 +26,7 @@ CREATE TABLE IF NOT EXISTS functional_units (
     surface DECIMAL(10,2) NOT NULL,
     surface_percentage DECIMAL(10,2) NOT NULL,
     tenant VARCHAR(25) DEFAULT NULL,
-    rent_value INT NOT NULL DEFAULT 0,
+    rent_value DECIMAL(10,2) NOT NULL DEFAULT 0.00,
     debt DECIMAL(10,2) DEFAULT 0,
     consortium INT NOT NULL,
     FOREIGN KEY (consortium) REFERENCES consortiums(id)
@@ -78,7 +78,7 @@ SELECT 1, '1A', 50.00, 5.00, 1
 );
 
 INSERT INTO functional_units (unit_number, unit_name, surface, surface_percentage, consortium, tenant, rent_value, debt)
-SELECT 2, '1B', 25.00, 2.50, 1, 'Carlos López', 350000, 700000
+SELECT 2, '1B', 25.00, 2.50, 1, 'Carlos López', 350000.00, 700000
     WHERE NOT EXISTS (
     SELECT 1 FROM functional_units WHERE unit_name = "1B" AND consortium = 1
 );
